ParticleContainer: Split
Split specializations for faster, more resource efficient compiles that
can be parallelized.
